scripts/mycelium_recipe_ops/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
PyNaCl>=1.5.0
